pyproject.toml
[project]
name = "cadastro"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Renato Chagas de Moraes", email = "renatochagas.m@gmail.com"},
]
dependencies = ["django>=4.2.18", "djangorestframework>=3.15.2"]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.scripts]
dev = "python manage.py runserver"
createsuperuser = "python manage.py createsuperuser"
migrate = "python manage.py migrate"
makemigrations = "python manage.py makemigrations"
